requirements.txt
paho-mqtt
pillow
numpy
requests
python-dateutil